requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
sqlalchemy
aiosqlite
python-multipart
email-validator
itsdangerous
openpyxl
pandas
weasyprint
slowapi
pytest
pytest-asyncio
pytest-cov
yagmail
